Add OnboardingInfo.lookup_county().
